Implement SETUP and TEARDOWN.
each test now include a null setup and teardown function pointers.
in case there exist a setup definition, it will be called before the main test execution while teardown will execute afterwards.

Setup and teardown provide a ctx variable of which the user is freely able to use for each test in the suite. This should help make the tests more useful and complete.

+ Implement NULL assertions