Reformat output to use less screen space and better expected/actual output.
maximum column length can be bad if it exceeds a certain point.
Currently, it's capped at 53, which i think is  ideal for a test framework. If truncation feels too strong, i may bump this number up a little.

This commit also centers the symbol in the suite print.

I also found it to be easier if the expected/actual are in new separate lines.