Add String assertion and allocation for the fail message.
Previously, the fail message would be truncated if it goes beyond 128 character. This is due to it being stack based implementation with a pipe between the parent and the forked process. This change allows the child process to allocate and send the length and the allocated buffer.

String assertion will print the full expected if any part of it does not match.