Refactor Internal String Buffer to use the same dynamic string as the internal.
This unifies the way we handle strings as well as allowing the header to use string_append to have a proper continuous string concatination.

This commit also changes the behavior of many internal macros as now they utilize LOG_TRACE_MSG() to log the file and line number. While allowing the user to make their own assertions by using LOG_TRACE_MSG and LOG_MSG.

This also introduces the ability to add any log message to any test, whether it's failing or passing (with exception to tests that crash).